Combine the Product and Reactant classes into the RecipeComponent class
