+++ calc/__init__.py
@@ -25,37 +25,27 @@ class Item:
 
 
 class RecipeComponent:
-    """This is the parent class for Reactants and Products. It is the generic representation of a
-    combination of an Item and a quantity.
+    """This is the class used to represent Reactants and Products. It is the generic representation
+    of a combination of an Item and a quantity. Whether something is a Product or Reactant depends
+    on whether it is one of the inputs or outputs of a recipe. This can change based on context.
 
     Attributes:
     item (Item): The Item object
     quantity (float): The quantity of the item. It has a float data type to allow precise
         calculations and probabilistic quantities.
+    component_type (str): The category of component (reactant or product)
     """
 
-    def __init__(self, item: Item, quantity: float):
+    def __init__(self, item: Item, quantity: float, component_type: str):
         self.item = item
         self.quantity = quantity
+        self.component_type = component_type
+        assert self.component_type in ["reactant", "product"] # Validate the component type
+
+    def info(self) -> str:
+        """Return the information about this recipe component."""
+        return f"{self.component_type.capitalize()}: {self.text()}"
 
     def text(self) -> str:
         """Return a string with the quantity and item name"""
         return f"{self.quantity} {self.item.name}"
-
-
-class Product(RecipeComponent):
-    """This is the class for Products, which are recipe components created using one or more
-    reactants.
-    """
-    def info(self) -> str:
-        """Return the information about this product."""
-        return f"Product: {self.text()}"
-
-
-class Reactant(RecipeComponent):
-    """This is the class for Reactants, which are recipe components used to create one or more
-    products.
-    """
-    def info(self) -> str:
-        """Return the information about this reactant."""
-        return f"Reactant: {self.text()}"
